move code up to core.

+++ src/Compilers/CSharp/Portable/Syntax/SyntaxListBuilder`1.cs
@@ -1,8 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using Microsoft.CodeAnalysis.CSharp.Symbols;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Text;
+using Microsoft.CodeAnalysis.Syntax;
 
 namespace Microsoft.CodeAnalysis.CSharp.Syntax
 {
@@ -69,7 +68,7 @@ namespace Microsoft.CodeAnalysis.CSharp.Syntax
 
         public bool Any(SyntaxKind kind)
         {
-            return _builder.Any(kind);
+            return _builder.Any((int)kind);
         }
 
         public SyntaxList<TNode> ToList()
